Update Projectile Messages and typos
Using $p does not work with throwables because it has already been transferred from the player. For a sling, the $p displays the name of the sling, not the intended rock.
This fix removes the <VOID> that will be shown, and also fixes some missile typos.
Note that $p still works with backstab and circle, because the player is still in possession of the weapon.